+++ dtwlib/dtwimage_p.h
@@ -138,7 +138,7 @@ public:
     DtwImagePrivate(DtwImage *q, const QSize& size);
 
     QImage makeImage() const;
-    QImage makeHighEnergyImage(int detailRatio) const;
+    QImage makeHighEnergyImage(float detailRatio) const;
 
     Seam findVerticalSeam() const;
     Seam findHorizontalSeam() const;
@@ -162,7 +162,7 @@ private:
 
     energy_t energy(int x, int y) const;
     void updateEnergy(index_t idx);
-    energy_t getThresholdEnergy(int ratio) const;
+    energy_t getThresholdEnergy(float ratio) const;
 
     energy_t dualGradientEnergy(int left, int rigth, int up, int down) const;
